fix: make new appId field works
