+++ prisma/schema.prisma
@@ -9,7 +9,7 @@ datasource db {
 
 model Message {
   id            Int        @id @default(autoincrement())
-  author        User       @relation(fields: [authorVaultId], references: [vaultId])
+  // author        User       @relation(fields: [authorVaultId], references: [vaultId])
   authorVaultId String
   text          String
   positionX     Float
@@ -30,14 +30,14 @@ model Whiteboard {
   authorVaultId String
   messages      Message[]
   curated       Boolean
-  joinedUsers   User[]    @relation("joinedWhiteboards")
-  favoriteUsers User[]    @relation("favoriteWhiteboards")
+  // joinedUsers   User[]    @relation("joinedWhiteboards")
+  // favoriteUsers User[]    @relation("favoriteWhiteboards")
 }
 
 model User {
-  vaultId             String       @id
-  createdWhiteboards  Whiteboard[]
-  joinedWhiteboards   Whiteboard[] @relation("joinedWhiteboards")
-  favoriteWhiteboards Whiteboard[] @relation("favoriteWhiteboards")
-  messages            Message[]
+  vaultId            String       @id
+  createdWhiteboards Whiteboard[]
+  // joinedWhiteboards   Whiteboard[] @relation("joinedWhiteboards")
+  // favoriteWhiteboards Whiteboard[] @relation("favoriteWhiteboards")
+  // messages            Message[]
 }
