+++ src/app/api/whiteboard/create/route.ts
@@ -1,8 +1,4 @@
 import { NextResponse } from "next/server";
-import { ethers } from "ethers";
-import { SiweMessage } from "siwe";
-import Fastify from "fastify";
-// import { FastifyInstance } from "fastify";
 
 import { prisma } from "../../db";
 import {
@@ -15,7 +11,6 @@ import {
 } from "@sismo-core/sismo-connect-server";
 import { verifyResponseMessage } from "../../common";
 import {
-  LOGO_BASE_64,
   MAX_CHARACTERS_WHITEBOARD_DESCRIPTION,
   MAX_CHARACTERS_WHITEBOARD_DESCRIPTION_MESSAGE,
   MAX_CHARACTERS_WHITEBOARD_NAME,
@@ -24,7 +19,7 @@ import {
   MAX_WHITEBOARD_GROUPS_MESSAGE,
   sismoConnectConfig,
 } from "@/app/configs/configs";
-import axios from "axios";
+import { getAppId } from "./utils";
 
 export const dynamic = "force-dynamic";
 
@@ -45,176 +40,6 @@ export async function POST(req: Request): Promise<NextResponse> {
     }
   } else {
     return NextResponse.json({ error: "No signed message" });
-  }
-  async function getAppId(whiteboardId: number): Promise<string> {
-    const wallet = ethers.Wallet.createRandom();
-    console.log("wallet.address", wallet.address);
-    // const res = await fetch("https://factory-api.sismo.io/apps/create",  {
-    //   method: "GET",
-    // });
-    const res = await axios.get(`https://factory-api.sismo.io/siwe/nonce`, {
-      withCredentials: true,
-    });
-    const nonce = res.data.nonce;
-    const nonceCookie = res.headers["set-cookie"]
-      ? res.headers["set-cookie"][0]
-      : "";
-    if (!nonceCookie) {
-      throw new Error("No nonceCookie");
-    }
-    console.log("nonceCookie", nonceCookie);
-    console.log("nonce", nonce);
-    const { message, signature } = await signSiweMessage(wallet, nonce);
-    // console.log("message", message);
-    // console.log("signature", signature);
-    // console.log(
-    //   "JSON.stringify({ message, signature })",
-    //   JSON.stringify({ message, signature })
-    // );
-    const bodyVerify = JSON.stringify({ message, signature });
-    console.log("bodyVerify", bodyVerify);
-    console.log("{ message, signature }", { message, signature });
-    // const fastify = Fastify({
-    //   bodyLimit: 2097152, // 2MB
-    //   ignoreTrailingSlash: true,
-    // });
-    // const cookies = {
-    //   address: fastify.signCookie(unsignedCookie.value ?? ""),
-    // };
-
-    // const config = {
-    //   method: "post",
-    //   url: "https://factory-api.sismo.io/siwe/verify",
-    //   headers: { "Content-Type": "application/json", Cookie: nonceCookie },
-    //   body: {
-    //     message: message,
-    //     signature: signature,
-    //   },
-    // };
-    // const responseVerify = await axios(config);
-    const responseVerify = await axios.post(
-      "https://factory-api.sismo.io/siwe/verify",
-      {
-        message: message,
-        signature: signature,
-      },
-      {
-        headers: { "Content-Type": "application/json", Cookie: nonceCookie },
-        withCredentials: true,
-      }
-    );
-    // const responseVerify = await fetch(
-    //   "https://factory-api.sismo.io/siwe/verify",
-    //   {
-    //     method: "POST",
-    //     headers: {
-    //       "Content-Type": "application/json",
-    //       Cookie: nonceCookie,
-    //     },
-    //     body: bodyVerify,
-    //   }
-    // );
-    // console.log("responseVerify", responseVerify);
-    // console.log("responseVerify.headers", responseVerify.headers);
-    // console.log("responseVerify.headers", responseVerify.headers);
-
-    // const cookies = [];
-
-    // for (const [key, value] of responseVerify.headers.entries()) {
-    //   if (key === "Set-Cookie") {
-    //     cookies.push(value);
-    //   }
-    // }
-    // console.log(
-    //   'res.headers["set-cookie"]',
-    //   responseVerify.headers["set-cookie"]
-    // );
-    // const headers: any = responseVerify.headers.entries();
-    // console.log('res.headers["set-cookie"]', headers);
-    // console.log("headers", headers);
-    // console.log('headers["cookies"]', headers["cookies"]);
-    console.log("responseVerify", responseVerify);
-    console.log("res.headers", responseVerify.headers);
-    console.log(
-      "res.headers['set-cookie']",
-      responseVerify.headers["set-cookie"]
-    );
-    const verifyCookie = responseVerify.headers["set-cookie"];
-    if (!verifyCookie) {
-      throw new Error("No verifyCookie");
-    }
-    //   ? res.headers["set-cookie"][0]
-    //   : "";
-    // if (!nonceCookie) {
-    //   throw new Error("No nonceCookie");
-    // }
-    const body = {
-      appInput: {
-        name: "zk-whiteboard-test-" + whiteboardId,
-        description:
-          "This is an app created by a user of zk-whiteboard.xyz. It allows to create a new whiteboard.",
-        authorizedDomains: ["zk-whiteboard.xyz"],
-        logoBase64: LOGO_BASE_64,
-        creatorId: wallet.address,
-      },
-    };
-    // const response = await fetch("https://factory-api.sismo.io/apps/create", {
-    //   method: "POST",
-    //   headers: {
-    //     "Content-Type": "application/json",
-    //     Cookie: nonceCookie,
-    //   },
-    //   body: JSON.stringify(body),
-    // });
-
-    const response = await axios.post(
-      "https://factory-api.sismo.io/apps/create",
-      body,
-      {
-        headers: { "Content-Type": "application/json", Cookie: verifyCookie },
-        withCredentials: true,
-      }
-    );
-
-    console.log("response", response);
-    console.log("response.data", response.data);
-    console.log("response.data.id", response.data.id);
-    // const responseJson = await response.json();
-    return await response.data.id;
-  }
-
-  async function signSiweMessage(
-    wallet: ethers.Wallet,
-    nonce: string
-  ): Promise<{
-    message: string;
-    signature: string;
-  }> {
-    const message = await createSiweMessage(
-      wallet.address,
-      "Sign in with Ethereum to the app.",
-      nonce
-    );
-    const signature = await wallet.signMessage(message);
-    return { message, signature };
-  }
-
-  async function createSiweMessage(
-    address: string,
-    statement: string,
-    nonce: string
-  ): Promise<string> {
-    const message = new SiweMessage({
-      domain: "factory.sismo.io",
-      address,
-      statement,
-      uri: "htpps://factory.sismo.io",
-      version: "1",
-      chainId: 1,
-      nonce,
-    });
-
-    return message.prepareMessage();
   }
 
   async function saveWhiteboard(
@@ -279,15 +104,6 @@ export async function POST(req: Request): Promise<NextResponse> {
       );
     }
 
-    // const whiteboard = await prisma.whiteboard.create({
-    //   data: {
-    //     name: signedMessage.message.name,
-    //     description: signedMessage.message.description,
-    //     groupIds: signedMessage.message.groupIds,
-    //     authorVaultId: vaultId,
-    //     curated: false,
-    //   },
-    // });
     const whiteboard = await prisma.whiteboard.create({
       data: {
         name: signedMessage.message.name,
@@ -295,16 +111,19 @@ export async function POST(req: Request): Promise<NextResponse> {
         groupIds: signedMessage.message.groupIds,
         curated: false,
         author: {
-          connect: { vaultId: vaultId }, // Assurez-vous que cette relation est correcte
+          connect: { vaultId: vaultId },
         },
         appId: "",
       },
     });
 
-    console.log("whiteboard", whiteboard);
-
     const sismoAppId = await getAppId(whiteboard.id);
+
     if (!sismoAppId) {
+      // Delete whiteboard if app creation failed
+      await prisma.whiteboard.delete({
+        where: { id: whiteboard.id },
+      });
       return NextResponse.json(
         {
           error: "App not created.",
@@ -312,6 +131,7 @@ export async function POST(req: Request): Promise<NextResponse> {
         { status: 500 }
       );
     }
+
     await prisma.whiteboard.update({
       where: { id: whiteboard.id },
       data: { appId: sismoAppId },
